raspberrypi-lcd-status/requirements.txt
psutil==6.1.1
RPLCD==1.3.1
smbus2==0.5.0
PyYAML==6.0.1
docker==5.0.3
kubernetes==22.6.0
